Added multiple items to Expensiveitems CSV, as well as adding multiple commands to the bot itself, and updating web requests to be asynchronous.
Commands added include:
Ping which makes the bot ping the specified role
NoPing which makes the bot not ping
Threshold which tells you the current priceratio
for it posting a message.
changethreshold(threshold) which will allow you to
change the threshold it should be a decimal IE: 0.3
would be 30% cheaper than exodus pricing.
currencyrefresh which refreshes the currency item
data frame based on an update to the csv.
itemrefresh which refreshes the Expensiveitems df
based on updates to that csv.
deleteitem(itemname,csv = False) which deletes an item from the
dataframe temporarily, it will be readded upon
running the itemrefresh command, if the csv command
is added as True then it will permanetely delete
the item from the csv.
additem(itemname,itemid,csvb = False) which adds an
item to the dataframe temporarily, it will be lost
among refresh of the dataframe. if csv is changed
to true then it will be added to the csv permanetely.
the permanent versions of delete / add item can only
be ran by the guild owner.